DEVGENRE-29 Lockup when using tool to warm up cache
In this fix I replace the usage of static HashSet that were not properly synchronized with ConcurrentHashMaps where I simply use the keySet and always set the same value (Boolean.TRUE). This is the same way the new Collections.newSetFromMap is implemented in JDK 1.6 but we couldn't use it since we still have to support JDK 1.5 in this branch.

There might be a performance impact since we are now properly locking the structures when adding or removing entries, but it should not have an impact on reading entries since they are not locked. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 44624)